Cross platform + Cleanup
If there were fewer PIDs than numberOfChromePIDsToWatch, the output got a little messy even though nothing bad happened but I fixed it to look pretty.

If we're on a Mac (and we have Applescript/osascript available) pop up a dialogue box. If we're not, just prompt the user in the terminal before timing out and killing the offensive page.